requirements.txt
albumentations==1.3.0
opencv-python==4.7.0.72
wandb==0.15.3
numpy==1.24.3
tqdm==4.65.0
segmentation-models-pytorch==0.3.2